- Navigation Graph added

+++ app/src/main/java/com/theshoremedia/database/helper/FactCheckHistoryDatabaseHelper.kt
@@ -40,23 +40,56 @@ class FactCheckHistoryDatabaseHelper {
         }
     }
 
-    fun getUnreadNews(callBack: (items: List<FactCheckHistoryModel>) -> Unit) {
+    fun getUnreadNews(
+        callBack: ((items: List<FactCheckHistoryModel>) -> Unit)
+    ) {
         handler?.post {
             if (AppDatabase.appDatabase == null) return@post
+
             callBack.invoke(
-                AppDatabase.appDatabase!!.factChecksHistoryDao()!!.getUnreadNews()
+                AppDatabase.appDatabase!!.factChecksHistoryDao()!!.getUnreadNews() ?: arrayListOf()
+            )
+        }
+    }
+
+    fun getFavouriteNews(
+        callBack: ((items: List<FactCheckHistoryModel>) -> Unit)
+    ) {
+        handler?.post {
+            if (AppDatabase.appDatabase == null) return@post
+
+            callBack.invoke(
+                AppDatabase.appDatabase!!.factChecksHistoryDao()!!.getFavouriteNews()
                     ?: arrayListOf()
             )
         }
     }
 
-    fun getFavouriteNews(callBack: (items: List<FactCheckHistoryModel>) -> Unit) {
+    fun getUnreadNewsLive(
+        owner: LifecycleOwner,
+        callBack: ((items: List<FactCheckHistoryModel>) -> Unit)
+    ) {
         handler?.post {
             if (AppDatabase.appDatabase == null) return@post
-            callBack.invoke(
-                AppDatabase.appDatabase!!.factChecksHistoryDao()!!.getFavouriteNews()
-                    ?: arrayListOf()
-            )
+            AppDatabase.appDatabase!!.factChecksHistoryDao()!!.getUnreadNewsLive().observe(
+                owner,
+                androidx.lifecycle.Observer {
+                    callBack.invoke(it ?: arrayListOf())
+                })
+        }
+    }
+
+    fun getFavouriteNews(
+        owner: LifecycleOwner,
+        callBack: ((items: List<FactCheckHistoryModel>) -> Unit)
+    ) {
+        handler?.post {
+            if (AppDatabase.appDatabase == null) return@post
+            AppDatabase.appDatabase!!.factChecksHistoryDao()!!.getFavouriteNewsLive().observe(
+                owner,
+                androidx.lifecycle.Observer {
+                    callBack.invoke(it ?: arrayListOf())
+                })
         }
     }
 
@@ -65,9 +98,17 @@ class FactCheckHistoryDatabaseHelper {
         val dummyData = ApplicationUtils.getDummyData(context)
         handler?.post {
             if (AppDatabase.appDatabase == null) return@post
+            if (AppDatabase.appDatabase!!.factChecksHistoryDao()?.getAllItems?.size != 0) return@post
             AppDatabase.appDatabase!!.factChecksHistoryDao()!!.deleteAll()
             AppDatabase.appDatabase!!.factChecksHistoryDao()!!.insertAll(dummyData)
         }
 
+    }
+
+    fun markAsFav(item: FactCheckHistoryModel) {
+        handler?.post {
+            if (AppDatabase.appDatabase == null) return@post
+            AppDatabase.appDatabase!!.factChecksHistoryDao()?.markedAsFav(item)
+        }
     }
 }
